added suspicious registry key

+++ cpu.h
@@ -33,7 +33,7 @@ double getCurrentValue() {
     GetProcessTimes(self, &ftime, &ftime, &fsys, &fuser);
     memcpy(&sys, &fsys, sizeof(FILETIME));
     memcpy(&user, &fuser, sizeof(FILETIME));
-    percent = (sys.QuadPart - lastSysCPU.QuadPart) +
+    percent = (double)(sys.QuadPart - lastSysCPU.QuadPart) +
         (user.QuadPart - lastUserCPU.QuadPart);
     percent /= (now.QuadPart - lastCPU.QuadPart);
     percent /= numProcessors;
